human-faces-count-model-api/requirements.txt
absl-py==2.3.1
addict==2.4.0
aioboto3==15.4.0
aiobotocore==2.25.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aioitertools==0.12.0
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.14.2
billiard==4.2.2
bleach==6.2.0
boto3==1.40.49
botocore==1.40.49
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.5
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
docopt==0.6.2
dotenv==0.9.9
easydict==1.13
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.1
facexlib==0.3.0
fastapi==0.119.1
fastjsonschema==2.21.2
filelock==3.19.1
filterpy==1.4.5
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
gevent==25.9.1
gfpgan==1.3.8
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.14.0
humanfriendly==10.0
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
insightface==0.7.3
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keras==2.14.0
kiwisolver==1.4.7
kombu==5.5.4
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.7.5
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.2.1
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.19.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.4.0
polars==1.34.0
polars-runtime-32==1.34.0
prettytable==3.16.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.1
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
stringzilla==4.2.1
sympy==1.14.0
tb-nightly==2.21.0a20251021
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2024.8.30
tinycss2==1.4.0
tomli==2.3.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.220
ultralytics-thop==2.0.17
urllib3==1.26.20
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wrapt==1.14.2
yapf==0.43.0
yarg==0.1.9
yarl==1.22.0
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
